fixed strong coupling but keep recursive render
