+++ package.json
@@ -13,6 +13,7 @@
 	"dependencies": {
 		"@emotion/react": "^11.11.1",
 		"@emotion/styled": "^11.11.0",
+		"@eurydos/structogram-component": "^1.0.2",
 		"@fontsource/fira-code": "^5.0.16",
 		"@fontsource/roboto": "^5.0.8",
 		"@mui/icons-material": "^5.14.19",
@@ -37,7 +38,7 @@
 		"file-saver": "^2.0.5",
 		"gh-pages": "^6.1.0",
 		"html-to-image": "^1.11.11",
-		"react-latex-next": "^2.2.0",
+		"react-latex-next": "^3.0.0",
 		"terser": "^5.26.0",
 		"typescript": "^5.2.2",
 		"vite": "^5.0.13",
