+++ src/App/App.tsx
@@ -1,13 +1,11 @@
-import { Fragment, FC } from "react";
 import {
 	CssBaseline,
 	ThemeProvider,
 	createTheme,
 } from "@mui/material";
-
-import { SnackbarProvider } from "notistack";
-
 import { LiveEditor } from "App/components/LiveEditor";
+import { SnackbarProvider } from "notistack";
+import { FC, Fragment } from "react";
 
 const themeDark = createTheme({
 	palette: {
