--- src/App/components/Diagram/index.ts
@@ -1,1 +1,0 @@
-export { Diagram } from "./Diagram";
