+++ src/main.tsx
@@ -1,14 +1,12 @@
-import React from "react";
-import ReactDOM from "react-dom/client";
-
-import { App } from "./App";
-
-import "katex/dist/katex.min.css";
 import "@fontsource/fira-code";
 import "@fontsource/roboto/300.css";
 import "@fontsource/roboto/400.css";
 import "@fontsource/roboto/500.css";
 import "@fontsource/roboto/700.css";
+import "katex/dist/katex.min.css";
+import React from "react";
+import ReactDOM from "react-dom/client";
+import { App } from "./App";
 
 ReactDOM.createRoot(
 	document.getElementById("root")!,
