+++ tsconfig.json
@@ -3,7 +3,7 @@
 		"baseUrl": "src",
 		"removeComments": true,
 
-		"target": "ES2020",
+		"target": "ES2022",
 		"useDefineForClassFields": true,
 		"lib": ["ES2020", "DOM", "DOM.Iterable"],
 		"module": "ESNext",
