+++ tsconfig.node.json
@@ -4,7 +4,8 @@
 		"skipLibCheck": true,
 		"module": "ESNext",
 		"moduleResolution": "bundler",
-		"allowSyntheticDefaultImports": true
+		"allowSyntheticDefaultImports": true,
+		"target": "es2022"
 	},
 	"include": ["vite.config.ts"]
 }
